refactor hindsight relabel fn as a data transformer
