+++ alf/algorithms/algorithm.py
@@ -1149,6 +1149,8 @@ class Algorithm(AlgorithmInterface):
             # in unroll().
             experience = self._add_batch_info(experience, batch_info)
             experience = self.transform_experience(experience)
+            # allow data_transformers to change batch_info
+            batch_info = experience.batch_info
             experience = self._clear_batch_info(experience)
         time_step = experience_to_time_step(experience)
         time_step, rollout_info = self.preprocess_experience(
