+++ alf/examples/ddpg_fetchreach.gin
@@ -13,7 +13,7 @@ create_environment.num_parallel_environments=38
 create_environment.env_name='FetchReach-v1'
 
 ObservationNormalizer.clipping=5.
-TrainerConfig.data_transformer_ctor=@ObservationNormalizer
+TrainerConfig.data_transformer_ctor=[@ObservationNormalizer, @HindsightExperienceTransformer]
 
 hidden_layers=(256, 256, 256)
