+++ alf/examples/her_fetchpush.gin
@@ -7,8 +7,7 @@ critic/CriticNetwork.observation_preprocessing_combiner=@NestConcat()
 critic/CriticNetwork.action_preprocessing_combiner=@NestConcat()
 
 ReplayBuffer.keep_episodic_info=True
-ReplayBuffer.postprocess_exp_fn=@hindsight_relabel_fn
-hindsight_relabel_fn.her_proportion=0.8
+HindsightExperienceTransformer.her_proportion=0.8
 
 DdpgAlgorithm.action_l2=0.05
